maintainers: Remove Paul and Kris
For now, remove Paul and Kris as they are not active in git at the moment.

Elevated git privileges come with responsibility of knowing git, being
comfortable with using it on a daily basis to help the project, and responsibly
merging code. It's a job, in effect, and not a privilege in itself.

I expect once the TSC forms, we may move to a per-repository maintainer
model.

Signed-off-by: Kyle Mestery <mestery@mestery.com>